Changed Id datatype from int to long for User

+++ moovi/src/main/java/com/dbms/project/moovi/data/repository/AdRecruiterRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 
-public interface AdRecruiterRepository extends CrudRepository<AdRecruiter, Integer>{
+public interface AdRecruiterRepository extends CrudRepository<AdRecruiter, Long>{
 
     @Query("SELECT a FROM AdRecruiter a WHERE a.username=:username")
     AdRecruiter findAdRecruiterByUsername(@Param("username") String u);
